Enable private connections for PostgreSQL
With this commit, we add private IPv4 and IPv6 ranges to the pg_hba.conf, thus
allowing connections from private subnet, which is should be faster as we don't
need to go out to public internet.
In practice, I didn't observe huge difference, because in one example I tried,
the route used by public communication was also pretty good. However, it is
possible to pick worse paths for public communication.