Add logging to bin/ci.
Previously E2E CI didn't log anything and was mostly silent except when
there was an exception. Having logs in its output is useful if something
goes wrong.

This PR logs strand state changes in standard output.

It logs when a strand is created, when its label changes, and when its
deleted.