Push a heartbeat event every 10 seconds
Our control plane is currently hosted on Heroku EU. Unfortunately, an
incident on Heroku (https://status.heroku.com/incidents/2590) impacted
our control plane too. We discovered this issue incidentally, as we
received no alert for it.

At present, we have two sources of alerts.

- StatusCake monitors our console and sends a page if it's down. We didn't
receive this alert because the web dyno was up.
- We receive apage  if a strand fails to reach the expected label before
its deadline. However, this logic is handled by respirate, and as
respirate was down, we didn't get a page.

This incident underscores the need to monitor respirate`s health as
well. Since respirate is just a worker without a web interface, services
like StatusCake can't ping it directly. Instead, we reverse monitoring
direction, also known as heartbeat monitoring, to check respirate's
health. respirate sends a heartbeat to a URL, and the listening service
notifies the relevant contacts if the heartbeat is missing.

StatusCake offers a "Push Monitoring" service for this purpose. If it
doesn't receive a push within a specified timeframe (60 seconds for us),
it creates a page.

`Prog::Heartbeat` ensures database connectivity before sending a
heartbeat. Additionally, if the HEARTBEAT_URL is set and no heartbeat
strand exists, one is created when respirate starts.