Only listen to SSH on GitHub runners
This patch is an expedient attempt to prevent malicious connections to
test workloads.

Some fellow staff members did some analysis of GitHub's runners and
the firewall situation, and with this patch, we end up much closer to
how GitHub's system works.

Substantially, this patch brings us to a similar situation at GitHub:
that is, connections initiated by the VM are permissive, and via
conntrack, allow return traffic.  But packets from the Internet that
attempt to establish new connections to the VM will be dropped.

We are more lax than GitHub in two respects:

1. As an exception, this patch allows SSH to listen.
2. We allow all ICMP, GitHub at least blocks ping.

(1) is for Clover to connect. In this case, we need a new connection
initiated from the Internet.

(2) Was mostly was spurred by IPv6, which at least needs neighbor
discovery protocol to work, but on consideration, other ICMP features,
such as Path MTU discovery, may also be useful.  This is the most
concise, useful definition.

The expediency relates it being more robust to do this packet
filtering on the VM Host, not allowing the guest to possibly gain root
and modify the nftables configuration.  But, that is more invasive,
so, I don't do it that way, yet.

== The Theory of Design

The GitHub runner image is a relatively complex beast, including use
of multiple containers for services, and implementing those containers
can have some number of extra network interfaces in the Vm that need
to communicate.

I'd also prefer to avoid an invasive patch to host behavior for this
expedient bit of firewall work.  So I try to do the necessary work
from within the VM hosting the GitHub runner.

The goal is to allow free communication within the Github Runner's VM,
including between virtual network interfaces and loopback connections,
without allowing Internet connections to be established to that
computer.

However, it should be able to initiate connections and receive return
traffic for them.  This is stateful, so, `conntrack` is going to be
used.

To square this circle, I elected to use something we know about
Internet traffic that's not true about other kinds of intra-Vm
traffic: it's always routed by the host, and carries the same source
mac address, from the host side of the TAP/TUN device.

== A Test Model

My model was like:

1. Start an ubicloud VM
2. Demonstrate that `nc -l $nat_inner_ipaddr 3000` can be connected to
   from home, using `echo oh snap | nc $nat_outer_ipaddr 3000`.  The
   listening side will print `oh snap`.
3. Use `sudo nft -f-` to install some rules like those seen in this
   patch, filling in the MAC address of the host side of the TAP/TUN
   device.  You can see the automated version in the code in this
   patch.
4. Flush conntrack information using `sudo conntrack -F` (otherwise,
   you can get path dependency between runs)
5. Demonstrate that `echo oh snap | nc $nat_outer_ipaddr 3000` doesn't
   work from my home ISP.
6. However, `echo oh snap | nc $nat_inner_ipaddr 3000` should on the
   instance, and it does not reject the traffic originating
   internally.
7. Also, `curl https://ubicloud.com`, and other Internet domains,
   should work.
8. Finally, check that new SSH connections can be established from the
   Internet, as Clover needs this to control the Runner's progression.