Protect postgres superuser password
Previously we processes superuser password just like any other information. For
example we kept it in the database unencrypted and passed it to VM along with
other configurations. That was OK because we don't have any managed PostgreSQL
instances running. However as a part of making PostgreSQL service production
ready, we should handle superuser password more carefuly, which is what this
commit does.