Add a binary in rhizome to reboot host.
`rhizome/host/bin/reboot-host` takes a boot id as an argument and if
matches the current boot id, then issues a reboot. Otherwise it prints
the current boot id to stdout.

This will be used by the control plane in a loop which will send reboot
messages to rhizome until we are sure reboot succeeded (i.e boot id
changed).