Retry reboot until boot id changes
Previously we had bunch of problems when detecting if reboot happened or
not. This was partly because `/sbin/reboot` is async and depending on
system status could return successfully or close connection.

This PR changes the approach by trying to send a reboot message to host
until boot id changes, or Vm::HostNexus's deadline arrives. Currently
deadline for HostNexus is 15 minutes.