Make the project mandatory in Nexus assemblies, even for development
We allowed the creation of resources without a project in development
mode to simplify the developers' workflow. However, even if we allow an
empty project, nexus will become stuck in subsequent steps. Furthermore,
we have a 'dev_project' helper in pry, eliminating the need to allow
nil project in assemble.