Make allocation robust to concurrent allocation_state changes
Because allocation uses two statements to perform allocation, it's
necessary to account for the scenario where an allocation outcome is
made using stale data, as in the following:

    1. Read & Assess

                              2. Situation change, COMMIT;

    3. Use assessment from (1) to allocate, COMMIT;

The third step in this chain makes the VM count on a VmHost in the
`draining` state non-monotonic.  This would be most noticeable in the
hypothetical presence of concurrent, long-lived transactions.  And the
Vm count being monotonic means that checking
`vm_host.vm_dataset.empty?` (spelled any number of ways) cannot be so
useful since it could be non-empty in the next moment.

This could have somewhat horrific ramifications, like moving forward
with deleting a host that was thought to be safely transitioned to the
`draining` state, but that gained a Vm at an inopportune time, right
before erasure.

The mechanism to account for this is checking an attribute on the
tuple to be updated as part of the `UPDATE` statement, the
`allocation_state`.  The resultant conditional write happens in one
command in SQL, and thus are subject to the interlock chain for
`UPDATE`, preventing undesired concurrency.

This defect was known about for a while, but some recent changes
people want to make it a bit more important to solve.  Mercifully, we
did not have to crash into the consequences of this in a live bug.