Handle a runner already exists edge case
When we generate a JIT config for a new runner using its UBID as the
name, GitHub creates an offline runner record and returns its ID. We
then save this runner ID to our GithubRunner record and initiate the
runner daemon with the JIT config. The runner switches from offline to
online when the script starts. If respirate exits for any reason such as
crash after sending the generate JIT config request but before saving
the runner_id, it will fail in the next iteration. This failure is due
to the runner already being created on the GitHub side, as indicated by
the following error message:

    409 - Already exists - A runner with the name *** already exists.

This case should be extremely rare, but we encountered it yesterday. If
we delete the existing runner, respirate will regenerate the JIT config
and proceed successfully.