Add a semaphore to defer vm destroy
In both production and development environments, I found it necessary to
delay the destroy of GitHub runner virtual machines. In the development,
I achieved this by commenting out the `incr_destroy` lines. In
production, I added a `sleep 100000` command into the workflow. However,
it's important to note that I cannot modify workflow files all the
times. Therefore, having a semaphore in place would be beneficial in
keeping virtual machines in a production environment for incident
investigations.

    # Prevent destroy
    vm.incr_prevent_destroy

    # Destroy deferred vm
    vm.strand.load.decr_prevent_destroy
    vm.incr_destroy

We are planning to add a script that creates a page when suspicious
network activity is detected. We need to delay the destroy of the VM to
investigate such incidents.