Inline Strand leasing mechanism
I was adding some logging that involved `@deleted`, when I realized
that the class method `def self.lease` was using an instance variable,
`@deleted`, that was never set in that context, and so never worked
properly.

Upon fixing it (by inlining), a look at the integrity check shows it
is nonsense, so, fix that too.

Alas.

Most of the time, I would use an accessor method to get a crash in
such a case, but seeing as this caused no problem and is terse and
efficient in a frequent, rarely changed code path, let's give me
another chance to get it right this time, and reduce the amount of
code, too.