Add Ubuntu 20.04 support for GitHub runners
We provide support for Ubuntu 22.04 on GitHub runners. The default
GitHub runners also support Ubuntu 20.04, and there is a packer template
available for this version. I've applied our custom packer provisioning
steps to Ubuntu 20.04, similar to what's done for 22.04, and was able to
generate a golden image for Ubuntu 20.04. The only difference is that
Ubuntu 20.04 doesn't come with `nftables`, which we use to block
incoming ports by default. However, I installed nftables during the
image generation process. To allow customers to choose the boot image,
I've added the boot image as a suffix to our current labels,
`ubicloud-standard-2-ubuntu-2004`. If the customer doesn't specify an
OS, Ubuntu 22.04 remains the default.