Use cp instead of mv to copy the GitHub runner script
When the `mv` command is executed, it removes the source file.
Therefore, if the label is rerun for any reason, it fails due to the
absence of the source file. The `cp` command, however, allows for
rerunning. Thus, we use it to ensure the `setup_environment` label is
idempotent.

Daniel recommended using the `-a` flag. It functions the same as the
`-RpP` options and preserves the structure and attributes of files.