Make the 'register_runner' label shorter and more idempotent
The 'register_runner' label is one of the most time-consuming labels. It
has low idempotency; it may fail or the respirate may exit at an
undesired line. We have encountered several incidents because of this.

To avoid storing the 'encoded_jit_config', we initiate an API call and
an SSH connection under the same label. However, we can separate the
part where we check the status of the runner-script. When the label has
a high line count, it becomes more difficult to debug because it's
challenging to locate the corresponding line. Additionally, we have a
status check at the 'wait' label, which I have now combined with the
other check.