Do not start the runner script if it's already running
In the "register_runner" label, if the respirate crashes after the
"./actions-runner/run.sh" script but before it hops, it fails in the
subsequent run. If it has already started, we can hop to the "wait"
label. The "wait" label determines the next course of action if the
service fails to start.