Introduce DNS Zones
This commit (along with subsequent commits) introduces internal-only DNS zones.
The intention is setting DNS zones and name servers up manually, but after they
are ready, it would be possible to add or remove records using the API provided
by DnsZone model.

We keep a full record of the dns records in our database, which is the primary
source of truth. This commit handles modifying DNS records in our database. In
the subsequent commits, we will also introduce logic to push these DNS records
to name servers.