Add MinIO cluster models
In this commit, we define the entity model for Minio clusters. MinIO has
a multi level cluster topology. Each cluster consists of multiple server
pools, each pool has multiple servers and each server has multiple
drives. In our entity model, they are represented in the following way;

MinioCluster
|
|-->MinioPool
    |
    |--> MinioServer <----> Vm
    |...
    |--> MinioServer <----> Vm
    			    |->VmStorageVolume (drive)
			    |->VmStorageVolume (drive)

The role of MinioPool may feel unnecessary or even weird, but it's
actually very essential. An object in minio is split to parts and stored
in a single MinioPool. The servers in a pool behave like they share all
the hardware they have. So then, you may ask, what good is it to have
MinioCluster? It's because of scalability. MinioPools do not allow to
add new servers or drives into the pool, however, we can always expand
the cluster by adding new pools. Minio will simply start storing the new
coming data in the most free pool. Furthermore, we can even decommision
a pool by moving its data into a newer pool. That's how MinIO mandates
the scalability operations to be handled.