Learn storage of /var/storage instead of /var
For GitHub runner hosts, we mount `/var/storage` from a separate btrfs
disk partition. As a result, `/var` does not include the size of
`/var/storage`, which has more storage. We manually update the
`total_storage_gib` and `available_storage_gib` of the VM host.

We can't check the size of /var/storage directly, because it might not
there for other hosts at this step. Until we unify our hosts with ext4
disks, we check /var/storage only if exists.