Set connection timeout for the heartbeat requests
The heartbeat prog attempts to connect to the database and send GET
requests to the heartbeat URL every 10 seconds. Although this process is
relatively straightforward, we've noticed that some heartbeat runs can
take 10-20 seconds at production, which could indicate a connection
issue.

Excon, by default, has a 60-second timeout. I reduced this to 2 seconds
to prevent excessive blocking of the respirate thread. If the respirate
cannot send a request to the heartbeat URL within 2 seconds, there may
be an issue either with our system or our 3rd-party monitoring service.