Register the vm provisioning deadline under a separate label
We have a known issue #718. If the label that we register deadline
fails, we can't persist the deadline information to the strand's stack.
When there is no space left, an exception is raised, preventing the vm
strand from saving the deadline. We don't receive pages due to slow
provisioning if the delay is caused by capacity issues.

I moved the deadline registration to a different label until we find an
elegant solution for issue #718.