Moves BillingRecord Finalize now() to DB
With this commit, we are moving the finalize Time.now() run completely
to the DB. This way, our updates while changing the billing record will
not be impacted by the timestamp difference in between controlplane and
the DB. We had 3 instances of VmPool reassignment failure because of
this issue. The problem is, the previous finalize logic depended on the
time setting of the controlplane process while the entity creation sets
the span.begin to db's time (now()). This happens only if the
controlplane's time setting is further than the DB's because of the
following scenario;
```
billing_record.update(begin_time, CONTROLPLANE_TIME.NOW())
billing_record.create(POSTGRES_TIME.now, null)
```
So, if PG is behind controlplane, the above code will have span
conflicting error.