Decrease MTU for all packets to 1400
We'd really rather not incur packet fragmentation from IPsec
encapsulation.  It is common to set MTUs on such encapsulated links or
routes to 1400.

However, this also affects non-encapsulated traffic (e.g. public
internet), which is unfortunate, as the lower overhead of the
customary 1500 bytes is generally accepted.

In principle, MTUs can be bound to routes via DHCP as well, avoiding
this de-optimization.  In practice, we'd have to check for support in
client DHCP implementations in common use, which is a lot more work
and prone to miss important platform targets.  In the interest of
expediency, this patch opts to be conservative to avoid fragmentation.

Support for this option was merged into systemd in 2019[^1].  Getting
the client to interpret the DHCP option was a live issue in practical
settings in 2018[^2].  This is not recent, but not ancient either.

[^1]: https://github.com/systemd/systemd/pull/14392
[^2]: https://unix.stackexchange.com/a/457577