Create root and server certificates for PG
We added a new label that uses Util.create_certificate for creating root and
server certificates. We are creating a root certificate per postgres server,
which saves us from storing a central and very powerful root certificate. On
the other hand, this means customers would need to use different root certs for
every server if they want to validate the server certificates. In the future
we might move to a root per organization (or whatever construct we would use
for aggregating projects).

Instead of creating our own CA, we also considered using one of the publicly
trusted CAs. However we decided against it because;
- Adding a network dependency reduces overall reliability.
- Any outage in the 3rd party system would cause outage in ours as well.
- It is slow to get certificates from these CAs, so it would increase our
provisioning times.
- It would also affect ease of development in negative way.

Some of these issues can be mitigated by creating a certificate pools but that
requires investment and would complicate the overall architecture.