Rotate PG certificates
Since creating and rotating certificates are quite similar, with few small
extra steps, we also got free certificate rotation. I even initially did the
initial certificate creation and certificate rotation in a single label, but
overall readability of the solution was not that good, so eventually decided
to have two separate labels with the cost of few extra lines.

We set the validity of the root certificates as 10 years and rotate them once
every 5 years. Each root certificate is used to sign new server certificates
between its 4th and 9th years. That means, depending on the the when the root
certs are downloaded, users can go 4 to 9 years without needing to re-download
root certificates. If they download the root certificates just after cluster
creation, they don't need to download the root certificates again for about 9
years. The worst case is when users download the root certificates just before
the rotation time. In this case, they would need to download the certificates
in 4 years.

Validity of the server certificate is 6 months. 1 month before the expiry of
the server certificate, we create a new server certificate using the older of
the two root certificates unless there is less than 1 year left to the expiry
of the older certificate. In that case, we use the newer one. Because in few
months, we would rotate the old certificate and we don't want to have a server
certificate that is signed by the almost expired root certificate.