Make the pool less eager to provision VMs
We realized there is an edge case that the eagerness of the pool to
create new VMs may cause starvation for some of the GithubRunners. Here
is a scenario;
1. We have a max capacity of 30 VMs
2. We have a pool with size 20
3. We get a request for 40 VMs very quickly. (Maybe a single request
or back to back runs)
4. We immediately make use of 20 VMs from the pool + we provision 10 new
VMs since we have the capacity. The other 10 VM request started to wait
for allocation. However, they fail, therefore, they exponentially
backoff.
5. Finally when the jobs start to finish and we have capacity to
provision VMs, the pool will try to provision immediately but the failed
GithubRunners will be too late to the party.

This PR solves this issue.