Implement a prog to redeliver failed GitHub webhook deliveries
GitHub does not resend failed webhook deliveries. These failures can
occur due to issues on our end or with a service we use, such as when
Cloudflare returns a 520 error or Heroku returns a 504 sometimes.

We need to pool deliveries and retry failed ones. The GitHub endpoint
returns a maximum of 100 items per page, requiring us to paginate until
we reach the expected date. If a failure arises, we might over-paginate
and hit GitHub's API rate limit. To mitigate this, I've added a maximum
page parameter. Given that we check every 10 minutes and each page
contains 100 items, I believe a limit of 2000 events is sufficient under
the current load. We log the last page number, so if it approaches 20,
we can consider adjusting the limit. Furthermore, an operator can
manually increase the max_page value at pry.

This program might generate additional load during an incident. For
instance, a recent issue with Heroku caused our web dyno to go down,
which will led this prog to repeatedly attempt to redeliver many events.
In such cases, we could consider deleting the strand. However, it
remains unclear to me how we might prevent such a scenario.