Hop to the correct label when reboot is requested.
In #674 we changed the label which initiated reboot from `reboot` to
`prep_reboot`. We do reboot in 2 cases:

1. When setting up host
2. On VmHost.incr_reboot

We had updated the label to be hopped at for case 1, but we had missed
updating the label for case 2. This PR fixes that.

To initiate reboot, you can do `vm_host.incr_reboot`.