Fixes IPv4 connectivity after host reboot
Issue #756 mentions the connectivity loss after host reboot for IPv4.
This is related to the ipv4 private subnet not being persistently added
to the tuntap* interface. In this PR, we are simply repopulating that
and getting into the rekeying sequence to also keep the private
networking connectivity. Because we also lost the ip xfrm state and
policy objects at the host reboot.