Simplify the logic for waiting on child strands
When we bud child strands, the parent strand reaps its children and
donates its running time to them. If it's a leaf, meaning no children
are left, it generally hops to the another label. I've attempted to
simplify this logic to avoid repetition across different locations.
The action a parent takes when it no longer has any children is more
common, so I' passed it as a block. The operation to process reaped
children is less common, so I passed it as a proc.