Move PostgreSQL's certificate configs to configure
We used to set ssl_cert_file and ssl_key_file during install_postgres. Since
these are config variables, it is better to set them during configure. On top
of that, when these are set during install_postgres, corresponding files are
not created yet (next label creates them), thus causing temporary failures.
These are not visible to customer as we didn't complete the provisioning yet,
but moving them to configure prevents that from happening anyway.