Add a fail helper to Clog that emits logs prior to an exception
Sometimes, we deliberately cause the code to fail to indicate that
something is wrong. However, when we raise an exception directly, it's
difficult to determine which model or strand is responsible for the
exception. The method Clog.fail first emits a log, then raises the
exception. This approach allows us to search the log with `thread:UBID`
and see the associated exceptions.