Only register deadline if allocation is successful
If we register deadlines during start, when there is a capacity problem we get
page storm. As a quick fix, we start to register deadlines only if allocation
is succeeded. This creates a blind spot for capacity issues, in a subsequent
PR, we will create a separate page to monitor capacity problems.