Moves boot image download to host_nexus for predefined boot images
With this commit, the predefined boot images are downloaded in the
host_nexus. The main reason I wanted to do this was that for a fresh
host, we had race condition while downloading images for the first
multiple concurrent VM provisioning. This way, the host comes with
images and we don't need to download them in conflicting ways.