Create a page if no capacity is left
Ideally we would create a page when capacity is low, but the purpose of this
commit it fixing the blind spot we added in cfa245c. In that commit, to prevent
page storms, we stopped generating one page per each unsuccessful allocation.
With this commit, we create one page for all failed allocations that are due to
lack of capacity. We do this by deduplicating pages based on tags "NoCapacity"
and the location of the requested VM.