Fix the eager loading issue in InvoiceGenerator
InvoiceGenerator retrieves all billing records for a given timeframe and
uses them for calculations. It requires the project associated with each
billing record. To avoid the N+1 query issue, we preload the project for
all billing records using `.eager(project: [:billing_info, :invoices])`.

Without eager loading, some projects have over 60k billing records,
which results in more than 60k queries to the database.

We use InvoiceGenerator to display current usage on the billing page.

I discovered that when we call `.map` directly on the dataset, as in
`active_billing_records.map`, it doesn't utilize eager loading. This
results in over 60k queries to the database.  To use eager loading, we
need to call the dataset with `.all`. I have updated this to
`active_billing_records.all.map`.