Use JSON-RPC directly to communicate with SPDK.
Previously we used SPDK's `rpc.py` to send commands to SPDK. This had
couple of drawbacks: (1) `rpc.py` is not normally installed with SPDK's
`make install` and we needed to manually add it to the package. (2) We
didn't parse the errors and treated all errors as equal.

This PR adds a simple JsonRpcClient class which is used by SpdkRpc to
communicate with SPDK.

I considered 2 existing json-rpc gems (jimson and json-rpc-client), but
when I tried them they weren't able to connect to unix sockets, and also
they didn't seem mature. Implementing a simple JsonRpc client was easy,
so I did that.

In addition to refactoring, this PR also fixes the issue that we failed
purge if SPDK objects didn't exist. After this PR, we don't raise an
error if a bdev/key/vhost controller to be delete doesn't exist.