Makes cloudinit function idempotent
We delete cloudinit_img before running mkdosfs command because it is not
idempotent. This causes a stuck provisioning in case a failure happens
in later stages of vm_setup. Especially because download_boot_image has
tendency to crash if concurrent image download is triggered. In that
case, the download_boot_image is actually idempotent but since mkdosfs
call is not, we get a stuck provisioning. We solve the issue by simply
removing the file and recreating.