Program to restart SPDK.
We will need to restart SPDK when we upgrade it. This patch makes that
possible by introducing `Prog::Storage::RestartSpdk`, which can be
invoked as follows:

```
st = vm_host.restart_spdk
st.run 300
```

Note that VMs aren't stopped when SPDK is being restarted. However, they
might see a big delay in I/O operations while the restart completes.
After SPDK goes down, cloud-hypervisor will try to connect to SPDK for
60 seconds before announcing a VM as failed. This timeout is hard-coded
in cloud-hypervisor, and is not configurable [1].

[1]
https://github.com/cloud-hypervisor/cloud-hypervisor/blob/ffb9a051c3bc9f1f417863f9635b105343b71653/virtio-devices/src/vhost_user/vu_common_ctrl.rs#L405