Emit stderr and other information via Clog
Quite a few times, we've found ourselves wishing that stderr was
logged from failed SSH commands.  I was first hesitant about the
prospect of what could end up in the logging system as a result.

But, after reviewing how we use `cmd`, and other opportunities for
data spilling (`SshError` also has stderr and stdout as fields), I
think the overall log risk is marginal at best, and the benefits
clear.

Moreover, I know that for very obscure issues, the stderr generated by
processes can be useful, reflecting failed syscalls or other bizarre
behavior (e.g. when a disk fails), so I think there is value in
defaulting to including stderr output.

One alternative design that I thought had promise was to store logs on
the Sshable host itself, thus harmonizing the log situation with uses
of `systemd-run` e.g. via daemonizer, but I found limitations in
simple ways to achieve this: we'd need to introduce a process that
copy output to two file descriptors (one for the log, one for Clover),
add extra system calls and process executions, and put another process
into the process tree.  Maybe these costs are acceptable, but they do
add a few milliseconds in a synchronous path, and extra memory commits
that can fail, for what is otherwise a basic action in this system.