Provide more storage for the larger GitHub runners
We configured our GitHub runner images to be 86GB, with ~16GB of free
space, excluding all installed software. This might suffice for small
tasks, but for heavy workload jobs that require powerful runners, it
might not be enough. We encountered workflows where this free space was
insufficient.

Additionally, larger GitHub runners offer more storage: 4vCPU provides
150GB, 8vCPU provides 300GB, and 16vCPU provides 600GB.

We initially opted for slightly lower storage figures. The rationale
behind our decision is that it's easier to scale up storage sizes as
required in the future, whereas reducing them would be more challenging.

    standard-2 runner     86GB
    standard-4 runner    150GB
    standard-8 runner    200GB
    standard-16 runner   300GB