Reset exponential backoff when hopping
Since `hop` is a success situation, it doesn't make sense to carry on
the exponential backoff `try` number.  Re-setting `try` to `0` makes
it consistent with `nap`'s behavior.

This was an oversight in 16bec8e26d4c9a398939f78a99c2fae1db657f41, in
which exponential backoff was added altogether.