Tighten up hash traversal in github webhook
Avoid re-traversing the hash with the same literal all the time,
instead binding the result to a variable.  Also, use `.fetch`, it
looks like the code is not written to cope with these fields being
`nil`, so let's convey this.

Some test fixtures needed updating, I checked the manual[^1] to verify
that the fields are "required," and thus always present.

[^1]: https://docs.github.com/en/webhooks/webhook-events-and-payloads?actionType=in_progress#workflow_job