Keep daily billing records for GitHub runners
GitHub runner resources are ephemeral, yet numerous. Currently, we
provision +3000 runners each day for some customers. Generating a
billing record for each would result in a monthly invoice with 90,000
line items. Burak proposed a solution, and we have decided to aggregate
billing records on a daily basis to reduce the number of entries.

We either generate a new billing record or augment the existing one for
the day, if it exists. The billing record table has an exclusion
constraint, which prevents the creation of multiple billing records for
the same day. If a thread encounters this constraint, it immediately
retries. I aim to finalize the billing rate promptly and don't record
the higher number for the used amount. The process retries up to 4
times, although I do not anticipate needing all four retries in a single
run.