Make the tag unique for only active pages
We implemented tag parts to avoid creating duplicate pages. However,
`Page.from_tag_parts` also returns previously resolved pages. This
implies that if we've already created and resolved a page, we cannot
create a new one with the same tag parts.

I discovered this issue when we encountered a capacity problem today and
didn't get a page. This was due to an existing resolved page with
`["NoCapacity", vm.location]` tag parts.

The same issue applies to deadline pages. If we have a resolved page
with `[id, prog, stack.first["deadline_target"]]` tag parts, we are
unable to generate a new page for the same target within the same prog.